Labs/BEW-1.2-Events-Homework/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
python-dotenv==0.15.0
SQLAlchemy==1.3.22